aegis_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
redis
pandas
polars
scikit-learn
xgboost
joblib
pydantic
pydantic-settings
python-multipart
httpx
pytest
pytest-asyncio
faker
python-jose[cryptography]